--- apps/ui/src/components/types.ts
@@ -1,4 +1,0 @@
-export interface SideBarProps {
-  isSchemaActive?: boolean;
-  isDashboard?: boolean;
-}
