+++ apps/ui/src/utils/create-tree-data.ts
@@ -4,20 +4,20 @@ export type RenderTree = Omit<Span, "__typename"> & {
   children: RenderTree[];
 };
 
-const sortSpansByStartTime = (a: RenderTree, b: RenderTree): number => {
+function sortSpansByStartTime(a: RenderTree, b: RenderTree): number {
   return Number(BigInt(a.startTimeUnixNano) - BigInt(b.startTimeUnixNano));
-};
+}
 
-const sortTreeDataRecursively = (tree: RenderTree[]): RenderTree[] => {
+function sortTreeDataRecursively(tree: RenderTree[]): RenderTree[] {
   for (const span of tree) {
     if (span.children.length > 0) {
       span.children = sortTreeDataRecursively(span.children);
     }
   }
   return tree.sort(sortSpansByStartTime);
-};
+}
 
-export const createTreeData = (spanArray: Span[]): RenderTree[] => {
+export function createTreeData(spanArray: Span[]): RenderTree[] {
   const treeData: RenderTree[] = [];
   const lookup: { [key: string]: RenderTree } = {};
 
@@ -51,4 +51,4 @@ export const createTreeData = (spanArray: Span[]): RenderTree[] => {
   });
 
   return sortTreeDataRecursively(treeData);
-};
+}
